Move MultiString to its own file
